+++ packages/config/src/configs/centrifuge.ts
@@ -1,0 +1,25 @@
+import {
+  BalanceBuilder,
+  ExtrinsicBuilder,
+  FeeBuilder,
+} from '@moonbeam-network/xcm-builder';
+import { cfg } from '../assets';
+import { centrifuge, moonbeam } from '../chains';
+import { AssetConfig } from '../types/AssetConfig';
+import { ChainConfig } from '../types/ChainConfig';
+
+export const centrifugeConfig = new ChainConfig({
+  assets: [
+    new AssetConfig({
+      asset: cfg,
+      balance: BalanceBuilder().system().account(),
+      destination: moonbeam,
+      destinationFee: {
+        amount: FeeBuilder().assetManager().assetTypeUnitsPerSecond(),
+        asset: cfg,
+      },
+      extrinsic: ExtrinsicBuilder().xTokens().transfer(),
+    }),
+  ],
+  chain: centrifuge,
+});