+++ packages/config/src/configs/index.ts
@@ -1,11 +1,13 @@
 import { ChainConfig } from '../types/ChainConfig';
 
 import { acalaConfig } from './acala';
+import { alphanetAssetHubConfig } from './alphanetAssetHub';
 import { alphanetRelayConfig } from './alphanetRelay';
 import { astarConfig } from './astar';
 import { bifrostKusamaConfig } from './bifrostKusama';
 import { bifrostPolkadotConfig } from './bifrostPolkadot';
 import { calamariConfig } from './calamari';
+import { centrifugeConfig } from './centrifuge';
 import { crustShadowConfig } from './crustShadow';
 import { darwiniaConfig } from './darwinia';
 import { darwiniaCrabConfig } from './darwiniaCrab';
@@ -19,6 +21,7 @@ import { karuraConfig } from './karura';
 import { khalaConfig } from './khala';
 import { kintsugiConfig } from './kintsugi';
 import { kusamaConfig } from './kusama';
+import { kusamaAssetHubConfig } from './kusamaAssetHub';
 import { litentryAlphanetConfig } from './litentryAlphanet';
 import { litmusConfig } from './litmus';
 import { mangataKusamaConfig } from './mangataKusama';
@@ -30,12 +33,11 @@ import { parallelConfig } from './parallel';
 import { parallelHeikoConfig } from './parallelHeiko';
 import { phalaConfig } from './phala';
 import { polkadotConfig } from './polkadot';
+import { polkadotAssetHubConfig } from './polkadotAssetHub';
 import { robonomicsConfig } from './robonomics';
 import { shidenConfig } from './shiden';
-import { statemineConfig } from './statemine';
-import { statemineAlphanetConfig } from './statemineAlphanet';
-import { statemintConfig } from './statemint';
 import { subsocialAlphanetConfig } from './subsocialAlphanet';
+import { turingAlphanetConfig } from './turingAlphanet';
 import { uniqueAlphaConfig } from './uniqueAlpha';
 
 export const chainsConfigList: ChainConfig[] = [
@@ -45,6 +47,7 @@ export const chainsConfigList: ChainConfig[] = [
   bifrostKusamaConfig,
   bifrostPolkadotConfig,
   calamariConfig,
+  centrifugeConfig,
   crustShadowConfig,
   darwiniaConfig,
   darwiniaCrabConfig,
@@ -71,10 +74,11 @@ export const chainsConfigList: ChainConfig[] = [
   polkadotConfig,
   robonomicsConfig,
   shidenConfig,
-  statemineAlphanetConfig,
-  statemineConfig,
-  statemintConfig,
+  alphanetAssetHubConfig,
+  kusamaAssetHubConfig,
+  polkadotAssetHubConfig,
   subsocialAlphanetConfig,
+  turingAlphanetConfig,
   uniqueAlphaConfig,
 ];