+++ packages/config/src/configs/polkadotAssetHub.ts
@@ -4,31 +4,31 @@ import {
   ExtrinsicBuilder,
   FeeBuilder,
 } from '@moonbeam-network/xcm-builder';
-import { tt1, unit } from '../assets';
-import { moonbaseAlpha, statemineAlphanet } from '../chains';
+import { dot, usdt } from '../assets';
+import { moonbeam, polkadotAssetHub } from '../chains';
 import { AssetConfig } from '../types/AssetConfig';
 import { ChainConfig } from '../types/ChainConfig';
 
-export const statemineAlphanetConfig = new ChainConfig({
+export const polkadotAssetHubConfig = new ChainConfig({
   assets: [
     new AssetConfig({
-      asset: tt1,
+      asset: usdt,
       balance: BalanceBuilder().assets().account(),
-      destination: moonbaseAlpha,
+      destination: moonbeam,
       destinationFee: {
         amount: FeeBuilder().assetManager().assetTypeUnitsPerSecond(),
-        asset: tt1,
+        asset: usdt,
       },
       extrinsic: ExtrinsicBuilder()
         .polkadotXcm()
         .limitedReserveTransferAssets()
         .X2(),
       fee: {
-        asset: unit,
+        asset: dot,
         balance: BalanceBuilder().system().account(),
       },
       min: AssetMinBuilder().assets().asset(),
     }),
   ],
-  chain: statemineAlphanet,
+  chain: polkadotAssetHub,
 });