+++ packages/builder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @moonbeam-network/xcm-builder
 
+## 1.0.4
+
+### Patch Changes
+
+- [#108](https://github.com/PureStake/xcm-sdk/pull/108) [`dcee04b`](https://github.com/PureStake/xcm-sdk/commit/dcee04be5df5c90027534a37e40e459ae0a14fa2) Thanks [@mmaurello](https://github.com/mmaurello)! - Change Statemint/Statemine names to Asset Hub
+
+## 1.0.3
+
+### Patch Changes
+
+- [#104](https://github.com/PureStake/xcm-sdk/pull/104) [`f1fc312`](https://github.com/PureStake/xcm-sdk/commit/f1fc312d77e8d579e40e48e24f3c139599bf74ca) Thanks [@mmaurello](https://github.com/mmaurello)! - Turing network integration with Moonbase Alpha
+
 ## 1.0.2
 
 ### Patch Changes